requirements.txt
requests
python_telegram_bot
